# design/openapi.yaml
openapi: 3.0.3
info:
  title: EventPulse API
  description: |
    A comprehensive event management and analytics platform built with Spring Boot. 
    EventPulse provides real-time event ingestion, processing, and analytics capabilities 
    with JWT authentication, WebSocket support, and detailed metrics.
  version: 1.0.0
  contact:
    name: EventPulse Team
    email: support@eventpulse.com
    url: https://eventpulse.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development Server
  - url: https://api.eventpulse.com
    description: Production Server
security:
  - bearerAuth: []
paths:
...
  /api/events:
    get:
      tags:
        - Events
      summary: Get Events
      description: Retrieve events with optional filtering by type, source, and time range
      parameters:
        - name: type
          in: query
          description: Filter events by type
          schema:
            type: string
            example: "login"
        - name: source
          in: query
          description: Filter events by source
          schema:
            type: string
            example: "user-service"
        - name: startTime
          in: query
          description: Start time for time range filter (ISO-8601 format)
          schema:
            type: string
            format: date-time
            example: "2023-12-01T00:00:00Z"
        - name: endTime
          in: query
          description: End time for time range filter (ISO-8601 format)
          schema:
            type: string
            format: date-time
            example: "2023-12-01T23:59:59Z"
      responses:
        '200':
          description: Events retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventResponse'
        '400':
          description: Invalid filter parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Events
      summary: Create Event
      description: Create a new event with source, type, and message. The timestamp is automatically set to current time.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequest'
            examples:
              login_event:
                summary: User login event
                value:
                  source: "user-service"
                  type: "login"
                  message: "User john.doe logged in successfully"
              error_event:
                summary: Error event
                value:
                  source: "payment-service"
                  type: "error"
                  message: "Payment processing failed for order #12345"
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/events/{id}:
    get:
      tags:
        - Events
      summary: Get Event by ID
      description: Retrieve a specific event by its unique identifier
      parameters:
        - name: id
          in: path
          required: true
          description: Event UUID
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: Event retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid UUID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Events
      summary: Delete Event
      description: Delete a specific event by its unique identifier
      parameters:
        - name: id
          in: path
          required: true
          description: Event UUID
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '204':
          description: Event deleted successfully
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid UUID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login endpoint
  schemas:
...
    EventRequest:
      type: object
      required:
        - source
        - type
        - message
      properties:
        source:
          type: string
          maxLength: 255
          description: Source of the event
          example: "user-service"
        type:
          type: string
          maxLength: 100
          description: Type of the event
          example: "login"
        message:
          type: string
          maxLength: 2000
          description: Event message
          example: "User john.doe logged in successfully"
    EventResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique event identifier
          example: "123e4567-e89b-12d3-a456-426614174000"
        source:
          type: string
          description: Source of the event
          example: "user-service"
        type:
          type: string
          description: Type of the event
          example: "login"
        message:
          type: string
          description: Event message
          example: "User john.doe logged in successfully"
        timestamp:
          type: string
          format: date-time
          description: Event timestamp
          example: "2023-12-01T10:30:00Z"
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2023-12-01T10:30:00Z"
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error type
          example: "Bad Request"
        message:
          type: string
          description: Error message
          example: "Invalid request data"
        details:
          type: object
          description: Additional error details
          additionalProperties:
            type: string
          example:
            source: "Event source is required"
            type: "Event type must not exceed 100 characters"